Send g-code to the printer.
Tested with Repetier + NanoDLP mods firmware.
Theoretically, it should work without large fixes required anywhere.

Right now, it always waits for "ok N" before sending a command. It
means that if a response is lost, it will hang. This is a serious
problem that must be addressed before releasing it to the real users.

See https://github.com/robodone/tasks/issues/17 that tracks this.